Clean up Kitty config and add 8-pane split
- Add split_8_panes.py kitten for 2x4 grid layout
- Remove old/unused split methods from kitty.conf
- Delete test files and obsolete scripts
- Keep only the working Python kitten approach

Both Cmd+4 (4 panes) and Cmd+8 (8 panes) now use the reliable
Python kitten method that preserves working directory.
