+++ kitty/kitty.conf
@@ -74,26 +74,14 @@ map cmd+d launch --location=vsplit --cwd=current
 map cmd+shift+d launch --location=hsplit --cwd=current
 
 # Cmd+4 = Split into 4 equally-sized panes (2x2 grid)
-# Fixed: Ensures all panes inherit the parent's working directory
-map cmd+4 combine : launch --location=hsplit --cwd=current : launch --location=vsplit --cwd=current : neighboring_window up : launch --location=vsplit --cwd=current : send_text all cd .\r
+map cmd+4 kitten split_4_panes.py
 
 # Cmd+8 = Split into 8 equally-sized panes (2x4 grid)
-# Original without delays for now
-map cmd+8 combine : launch --location=hsplit --cwd=current : launch --location=vsplit --cwd=current : neighboring_window up : launch --location=vsplit --cwd=current : launch --location=vsplit --cwd=current : neighboring_window down : launch --location=vsplit --cwd=current : neighboring_window left : neighboring_window left : launch --location=vsplit --cwd=current : neighboring_window up : launch --location=vsplit --cwd=current
+map cmd+8 kitten split_8_panes.py
 
 # Cmd+R = Refresh theme in all panes (run cd . in each)
 map cmd+r combine : send_text all cd .\r
 
-# Cmd+5 = Alternative 4-pane split using remote control (more reliable)
-# This uses @ commands which better preserve the working directory
-map cmd+5 launch --type=overlay sh -c 'cwd=$(pwd); kitty @ launch --type=window --location=hsplit --cwd="$cwd" && kitty @ launch --type=window --location=vsplit --cwd="$cwd" && kitty @ focus-window --match neighbor:up && kitty @ launch --type=window --location=vsplit --cwd="$cwd"'
-
-# Cmd+6 = 4-pane split using custom kitten (most reliable)
-# This requires split_4_panes.py to be in ~/.config/kitty/
-map cmd+6 kitten split_4_panes.py
-
-# Cmd+7 = Test kitten
-map cmd+7 kitten test_kitten.py
 
 # Cmd+Enter = Make current window full screen (and back)
 map cmd+enter toggle_layout stack
