--- kitty/split-8-panes.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-# Split current kitty window into 8 panes (2x4 grid)
-# Note: This script requires running from within Kitty terminal
-
-echo "This functionality is built into Kitty config."
-echo "Use Cmd+8 to create 8-pane layout."
-echo "Remote control setup would be needed for CLI usage."
